feat(#2148): Moce DummyCentral into FakeMaven class
