+++ eo-maven-plugin/pom.xml
@@ -454,20 +454,20 @@ SOFTWARE.
           <plugin>
             <groupId>com.github.volodya-lombrozo</groupId>
             <artifactId>jtcop-maven-plugin</artifactId>
-            <configuration>
-              <!--
-                @todo #2144:30min Enable RuleAllTestsHaveProductionClass rule.
-                  It is disabled in the whole eo-maven-plugin module because
-                  it requires lots of fixes simultaneously. We have to enable
-                  that rule and apply all required changes in order to make
-                  all jtcop checks to pass.
-              -->
-              <exclusions>
-                <exclusion>
-                  JTCOP.RuleAllTestsHaveProductionClass
-                </exclusion>
-              </exclusions>
-            </configuration>
+<!--            <configuration>-->
+<!--              &lt;!&ndash;-->
+<!--                @todo #2144:30min Enable RuleAllTestsHaveProductionClass rule.-->
+<!--                  It is disabled in the whole eo-maven-plugin module because-->
+<!--                  it requires lots of fixes simultaneously. We have to enable-->
+<!--                  that rule and apply all required changes in order to make-->
+<!--                  all jtcop checks to pass.-->
+<!--              &ndash;&gt;-->
+<!--              <exclusions>-->
+<!--                <exclusion>-->
+<!--                  JTCOP.RuleAllTestsHaveProductionClass-->
+<!--                </exclusion>-->
+<!--              </exclusions>-->
+<!--            </configuration>-->
           </plugin>
         </plugins>
       </build>
